+++ engine/server/sv_game.cpp
@@ -33,7 +33,7 @@ botlib_export_t	*botlib_export;
 int	SV_NumForGentity( sharedEntity_t *ent ) {
 	int		num;
 
-	num = ( (byte *)ent - (byte *)sv.gentities ) / sv.gentitySize;
+	num = (int)( ( (byte *)ent - (byte *)sv.gentities ) / sv.gentitySize );
 
 	return num;
 }
@@ -71,7 +71,7 @@ svEntity_t	*SV_SvEntityForGentity( sharedEntity_t *gEnt ) {
 sharedEntity_t *SV_GEntityForSvEntity( svEntity_t *svEnt ) {
 	int		num;
 
-	num = svEnt - sv.svEntities;
+	num = (int)( svEnt - sv.svEntities );
 	return SV_GentityNum( num );
 }
 
@@ -348,7 +348,7 @@ int GameImport_FS_GetFileList( const char * path, const char * extension, char *
 
 }
 int GameImport_FS_Seek( int f, int64_t offset, int origin ) {
-	return FS_VM_SeekFile( f, offset, (fsOrigin_t)origin, H_QAGAME );
+	return FS_VM_SeekFile( f, (long)offset, (fsOrigin_t)origin, H_QAGAME );
 
 }
 void GameImport_SendConsoleCommand( int exec_when, const char * text ) {
