+++ game/ui/ui_serverinfo.cpp
@@ -140,8 +140,8 @@ static void ServerInfo_MenuDraw( void )
 
 		Q_strcat( key, MAX_INFO_KEY, ":" ); 
 
-		UI_DrawString(SCREEN_WIDTH*0.50f - 8,y,key,UI_RIGHT|UI_SMALLFONT,color_red);
-		UI_DrawString(SCREEN_WIDTH*0.50f + 8,y,value,UI_LEFT|UI_SMALLFONT,text_color_normal);
+		UI_DrawString((int)( SCREEN_WIDTH*0.50f - 8 ),y,key,UI_RIGHT|UI_SMALLFONT,color_red);
+		UI_DrawString((int)( SCREEN_WIDTH*0.50f + 8 ),y,value,UI_LEFT|UI_SMALLFONT,text_color_normal);
 
 		y += SMALLCHAR_HEIGHT;
 	}
